move all context related components into src/context
+ bugfix the "Sign out" on iOS
+ move the reboot logic in Reboot provider
