+++ src/components/CurrencyUnitInput.js
@@ -4,7 +4,7 @@ import type { Unit } from "@ledgerhq/currencies";
 import { View, StyleSheet, TextInput } from "react-native";
 import { formatCurrencyUnit, parseCurrencyUnit } from "@ledgerhq/currencies";
 import LText, { getFontStyle } from "../components/LText";
-import { withLocale } from "./LocaleContext";
+import { withLocale } from "../context/Locale";
 
 const initialValueStringFromProps = props =>
   props.value === 0
